refactor: Removed NativeUserRepository
